+++ GameLib.Plugin/GameLib.Plugin.Gog/GogGameFactory.cs
@@ -23,7 +23,7 @@ internal static class GogGameFactory
         return regKey.GetSubKeyNames()
             .AsParallel()
             .WithCancellation(cancellationToken)
-            .Select(gameId => LoadFromRegistry(launcher.ExecutablePath, gameId))
+            .Select(gameId => LoadFromRegistry(launcher, gameId))
             .Where(game => game is not null)
             .Select(game => AddLauncherId(launcher, game!))
             .ToList()!;
@@ -41,7 +41,7 @@ internal static class GogGameFactory
     /// <summary>
     /// Load the GoG game registry entry into a <see cref="GogGame"/> object
     /// </summary>
-    private static GogGame? LoadFromRegistry(string launcherExecutable, string gameId)
+    private static GogGame? LoadFromRegistry(ILauncher launcher, string gameId)
     {
         using var regKey = RegistryUtil.GetKey(RegistryHive.LocalMachine, $@"SOFTWARE\GOG.com\Games\{gameId}");
         if (regKey is null)
@@ -53,8 +53,7 @@ internal static class GogGameFactory
         {
             Id = (string)regKey.GetValue("gameID", string.Empty)!,
             Name = (string)regKey.GetValue("gameName", string.Empty)!,
-            ExecutablePath = (string)regKey.GetValue("exe", string.Empty)!,
-            Executable = (string)regKey.GetValue("exeFile", string.Empty)!,
+            Executable = (string)regKey.GetValue("exe", string.Empty)!,
             WorkingDir = (string)regKey.GetValue("workingDir", string.Empty)!,
             InstallDate = DateTime.TryParseExact(
                 (string)regKey.GetValue("INSTALLDATE", string.Empty)!, "yyyy-MM-dd HH:mm:ss",
@@ -83,8 +82,8 @@ internal static class GogGameFactory
             return null;
         }
 
-        game.InstallDir = Path.GetDirectoryName(game.ExecutablePath) ?? string.Empty;
-        game.LaunchString = $"\"{launcherExecutable}\" /command=runGame /gameId={game.Id}";
+        game.InstallDir = Path.GetDirectoryName(game.Executable) ?? string.Empty;
+        game.LaunchString = $"\"{launcher.Executable}\" /command=runGame /gameId={game.Id}";
         if (!string.IsNullOrEmpty(game.WorkingDir))
         {
             game.LaunchString += $" /path=\"{game.WorkingDir}\"";