+++ GameLib.Plugin/GameLib.Plugin.Gog/GogLauncher.cs
@@ -33,39 +33,35 @@ public class GogLauncher : ILauncher
 
     public string InstallDir { get; private set; } = string.Empty;
 
-    public string ExecutablePath { get; private set; } = string.Empty;
-
     public string Executable { get; private set; } = string.Empty;
 
-    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(Executable);
 
     public IEnumerable<IGame> Games { get; private set; } = Enumerable.Empty<IGame>();
 
     public void Refresh(CancellationToken cancellationToken = default)
     {
-        ExecutablePath = string.Empty;
         Executable = string.Empty;
         InstallDir = string.Empty;
         IsInstalled = false;
         Games = Enumerable.Empty<IGame>();
 
-        ExecutablePath = GetExecutable() ?? string.Empty;
-        if (!string.IsNullOrEmpty(ExecutablePath))
+        Executable = GetExecutable() ?? string.Empty;
+        if (!string.IsNullOrEmpty(Executable))
         {
-            Executable = Path.GetFileName(ExecutablePath);
-            InstallDir = Path.GetDirectoryName(ExecutablePath) ?? string.Empty;
-            IsInstalled = File.Exists(ExecutablePath);
+            InstallDir = Path.GetDirectoryName(Executable) ?? string.Empty;
+            IsInstalled = File.Exists(Executable);
             Games = GogGameFactory.GetGames(this, cancellationToken);
         }
     }
 
-    public bool Start() => IsRunning || ProcessUtil.StartProcess(ExecutablePath);
+    public bool Start() => IsRunning || ProcessUtil.StartProcess(Executable);
 
     public void Stop()
     {
         if (IsRunning)
         {
-            Process.Start(ExecutablePath, "/command=shutdown");
+            Process.Start(Executable, "/command=shutdown");
         }
     }
     #endregion