+++ GameLib.Plugin/GameLib.Plugin.Origin/Model/OriginGame.cs
@@ -11,9 +11,8 @@ public class OriginGame : IGame
     public Guid LauncherId { get; internal set; } = Guid.Empty;
     public string Name { get; internal set; } = string.Empty;
     public string InstallDir { get; internal set; } = string.Empty;
-    public string ExecutablePath { get; internal set; } = string.Empty;
     public string Executable { get; internal set; } = string.Empty;
-    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(Executable);
     public string WorkingDir { get; internal set; } = string.Empty;
     public string LaunchString { get; internal set; } = string.Empty;
     public DateTime InstallDate { get; internal set; } = DateTime.MinValue;