+++ GameLib.Plugin/GameLib.Plugin.Steam/SteamLauncher.cs
@@ -30,43 +30,39 @@ public class SteamLauncher : ILauncher
 
     public bool IsInstalled { get; private set; }
 
-    public bool IsRunning => ProcessUtil.IsProcessRunning(ExecutablePath);
+    public bool IsRunning => ProcessUtil.IsProcessRunning(Executable);
 
     public string InstallDir { get; private set; } = string.Empty;
 
-    public string ExecutablePath { get; private set; } = string.Empty;
-
     public string Executable { get; private set; } = string.Empty;
 
-    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(Executable);
 
     public IEnumerable<IGame> Games { get; private set; } = Enumerable.Empty<IGame>();
 
-    public bool Start() => IsRunning || ProcessUtil.StartProcess(ExecutablePath);
+    public bool Start() => IsRunning || ProcessUtil.StartProcess(Executable);
 
     public void Stop()
     {
         if (IsRunning)
         {
-            Process.Start(ExecutablePath, "-shutdown");
+            Process.Start(Executable, "-shutdown");
         }
     }
 
     public void Refresh(CancellationToken cancellationToken = default)
     {
-        ExecutablePath = string.Empty;
         Executable = string.Empty;
         InstallDir = string.Empty;
         IsInstalled = false;
         Libraries = Enumerable.Empty<SteamLibrary>();
         Games = Enumerable.Empty<IGame>();
 
-        ExecutablePath = GetExecutable() ?? string.Empty;
-        if (!string.IsNullOrEmpty(ExecutablePath))
+        Executable = GetExecutable() ?? string.Empty;
+        if (!string.IsNullOrEmpty(Executable))
         {
-            Executable = Path.GetFileName(ExecutablePath);
-            InstallDir = Path.GetDirectoryName(ExecutablePath) ?? string.Empty;
-            IsInstalled = File.Exists(ExecutablePath);
+            InstallDir = Path.GetDirectoryName(Executable) ?? string.Empty;
+            IsInstalled = File.Exists(Executable);
             Libraries = SteamLibraryFactory.GetLibraries(InstallDir);
             Games = SteamGameFactory.GetGames(this, Libraries);
         }